+++ R/playing_cards.r
@@ -62,7 +62,7 @@ pcbj_bavarian_pattern <- function(
 	back_notes <- paste(back_notes, collapse = "\n") |> marquee::marquee_glue(.trim = FALSE)
 	back <- marquee::marquee_grob(
 		back_notes,
-		style = credits_style("poker", color = text_col),
+		style = sbgjackets_style("poker", color = text_col),
 		width = unit(BAVARIAN_WIDTH, "in"),
 		x = unit(1 / 8 + BAVARIAN_DELTA, "in"),
 		y = unit(1, "npc") - unit(1 / 8, "in"),
@@ -154,7 +154,7 @@ pcbj_castilian_pattern <- function(
 	back_notes <- paste(back_notes, collapse = "\n") |> marquee::marquee_glue(.trim = FALSE)
 	back <- marquee::marquee_grob(
 		back_notes,
-		style = credits_style("poker", color = text_col),
+		style = sbgjackets_style("poker", color = text_col),
 		width = unit(pnpmisc:::JACKET_POKER_FRONT_WIDTH, "in"),
 		x = unit(1 / 8, "in"),
 		y = unit(1, "npc") - unit(1 / 8, "in")
@@ -236,7 +236,7 @@ pcbj_english_pattern <- function(
 	back_notes <- paste(back_notes, collapse = "\n") |> marquee::marquee_glue(.trim = FALSE)
 	back <- marquee::marquee_grob(
 		back_notes,
-		style = credits_style("poker", color = text_col),
+		style = sbgjackets_style("poker", color = text_col),
 		width = unit(pnpmisc:::JACKET_POKER_FRONT_WIDTH, "in"),
 		x = unit(1 / 8, "in"),
 		y = unit(1, "npc") - unit(1 / 8, "in")